+++ app/build.gradle.kts
@@ -67,6 +67,6 @@ dependencies {
     debugImplementation(libs.androidx.ui.tooling)
     debugImplementation(libs.androidx.ui.test.manifest)
     implementation("androidx.lifecycle:lifecycle-viewmodel-compose:2.6.1")
-    implementation("androidx.core:core-ktx:2.2.0")
+//    implementation("androidx.core:core-ktx:2.2.0")
     implementation("androidx.work:work-runtime-ktx:2.7.1")
 }
